Organise: Fixtures files into respective production + theatre directories
