+++ spec/fixtures/theatres/data-list.json
@@ -1,0 +1,7 @@
+{
+	"theatres": [
+		{
+			"id": 1
+		}
+	]
+}
