+++ spec/fixtures/theatres/data-with-errors.json
