+++ spec/fixtures/theatres/instance.json
